+++ Instagram/Instagram/Network/Base/NetworkBase.swift
@@ -1,0 +1,38 @@
+//
+//  NetworkBase.swift
+//  Instagram
+//
+//  Created by 김영인 on 2022/05/14.
+//
+
+import Foundation
+
+class NetworkBase {
+    
+    static func judgeStatus<T: Codable>(by statusCode: Int, _ data: Data, _ type: T.Type) -> NetworkResult<Any> {
+        let decoder = JSONDecoder()
+        guard let decodedData = try? decoder.decode(BaseResponse<T>.self, from: data)
+        else { return .pathErr }
+        
+        switch statusCode {
+        case 200 ..< 300: return .success(decodedData)
+        case 401 ..< 500: return .requestErr(decodedData)
+        case 500: return .serverErr
+        default: return .networkFail
+        }
+    }
+    
+    static func judgeOtherStatus<T: Codable>(by statusCode: Int, _ data: Data, _ type: T.Type) -> NetworkResult<Any> {
+        let decoder = JSONDecoder()
+        guard let decodedData = try? decoder.decode(type.self, from: data)
+        else { return .pathErr }
+        
+        switch statusCode {
+        case 200 ..< 300: return .success(decodedData)
+        case 401 ..< 500: return .requestErr(decodedData)
+        case 500: return .serverErr
+        default: return .networkFail
+        }
+    }
+    
+}
