+++ Instagram/Instagram/Source/Home/HomeViewController.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import SkeletonView
 
 class HomeViewController: UIViewController {
